renamed package to weather, and classes to names indiciative of what they do, stationlocator probably will need to be renamed
