+++ derby.log
@@ -1,5 +1,5 @@
 ----------------------------------------------------------------
-Fri Aug 02 11:16:31 CDT 2019D001 : [0] The Apache Software Foundation - Apache Derby - 10.15.1.3 - (1853019), [1] a816c00e-016c-531e-4ff3-000004a65df0, [2] C:\Users\peter.hydesmith\Documents\projects\weather-projects\weather\locales, [3] jdk.internal.loader.ClassLoaders$AppClassLoader@5e9f23b4, [4] file:/C:/Apache/db-derby-10.15.1.3-bin/lib/derby.jar
+Fri Aug 02 11:51:34 CDT 2019D001 : [0] The Apache Software Foundation - Apache Derby - 10.15.1.3 - (1853019), [1] a816c00e-016c-533e-68cc-000004ab8000, [2] C:\Users\peter.hydesmith\Documents\projects\weather-projects\weather\locales, [3] jdk.internal.loader.ClassLoaders$AppClassLoader@5e9f23b4, [4] file:/C:/Apache/db-derby-10.15.1.3-bin/lib/derby.jar
 java.vendor=Oracle Corporation
 java.runtime.version=12.0.1+12
 user.dir=C:\Users\peter.hydesmith\Documents\projects\weather-projects\weather
